New way of rendering osmc line color
Now all colors are near the highway - no more three rows
of color on each side - just one row on each side.
All colors can be renderend at the same time. It is not
so pretty as it used to be, but it works.
Also bicycle track lines are a bit improved.